Supply senders env in enif_send if done by a scheduler thread
Env must be present when invoking enif_send from a scheduler thread.

Some workarounds for things that have changed in the nif implemenation.
It might be a good idea to keep the 'extra' functionality
in the helper library 'clu', and keep 'cl' as close to the spec as
possible.

Minor Mac Makefile fixes

I messed up when converting to windows cl compiler
And cc_subdiv now compiles on mac, the display isn't that great on these
drivers I have on my mac-mini.  Everything is blended for some reason

+++ examples/cc_subdiv.cl
@@ -40,7 +40,7 @@ __kernel void gen_faces(
       center.xyz += VsIn[FsIn[fi.start+i]].xyz;
   }
 
-  center /= i;
+  center /= (float) i;
   // Create new center vertex
   const uint ov_id = noVs + face_id;
   center.w = fi.len*4.0;  // Valance = faceVs and hard_edge count = 0 (Valance << 2)
@@ -276,7 +276,7 @@ __kernel void move_verts(
     VsIn[v_id] = v_out;
   } else if(hc == 2) {
     v_out += v_in * 6.0;
-    v_out *= 1.0/8.0;
+    v_out *= 1.0f/8.0f;
     v_out.w = v_in.w;
     VsOut[v_id] = v_out;
     VsIn[v_id] = v_out;
